requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
ruff
scikit-learn
tqdm
typer
torch
torchvision
torchaudio
matplotlib
datasets
torchvision.models
Pillow
torch.nn
torch.optim
numpy
opencv-python
shap
transformers
-e .
